brain/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
jsonschema==4.20.0
google-generativeai
python-dotenv